+++ src/icalendar/cal/__init__.py
@@ -5,8 +5,8 @@ These are the defined components.
 """
 
 from .alarm import Alarm
-from .avaiable import Available
 from .availability import Availability
+from .available import Available
 from .calendar import Calendar
 from .component import Component
 from .component_factory import ComponentFactory
